[feat] Remove sorted input prior for auto angular resolution computation
Before this commit, the auto angular resolution computation relied on the assumption
that each scanline was already sorted by increasing timestamp/azimuth.
This implied a very specific organization of input cloud.

Now, each scanline is sorted, so this assumption is not needed anymore.
The only relying prior is that all points must be expressed in local lidar sensor
coordinates system (centered on (0, 0, 0) and spinning along -Z axis).