[feat] Create function to get config param with a string name
Create GetConfig function to get a parameter with his name.
To do it, I travel in the list of parameter and get the value associated
to the name.

Note: It take a lot of reseach in the source code of dynamic_reconfigure and
I can guarantee that it doesn't exist an more efficient way to do it.