[ROS2][feat] Autocompute Rpm and DeviceId
This commit removes Rpm and DeviceId as parameter for the user to enter.

It deduces DeviceId from the PCLHeader "frame_id" attribute.
It estimates Rpm thanks to the "stamp" attribute of the PCLHeader.
As Rpm is the difference between the time stamp of a current frame and the time stamp of its previous one,
the first frame has to be ignored (we can't access Rpm at first frame so time estimation will fail).
It checks the validity of the estimated Rpm according to a const vector of possible frequency values of the LiDAR.