[ROS2][feat] Separate static and moving objects in a cluster
For one big cluster in the current fov, check its standard deviation of pixels time.
If the value is large, check whether or not the cluster should be separated.

To do so, the cluster pixels are divided into two groups (recent and old).
If there are old pixels which split the cluster, they will be removed and
various subclusters will be created.